rag_retrieval_service/requirements.txt
functions-framework
google-cloud-firestore
vertexai
